WIP: develop CookieConsentBanner w/ Google Analytics
